Move hosts and etcd cluster spec packages out of protokube
pkg/dns/hosts (the /etc/hosts writer used by none-DNS clusters) and
pkg/etcd (etcd cluster spec parsing from volume tags) are still used
outside protokube, so relocate them ahead of the protokube removal.
protokube temporarily imports the new locations.
