+++ pkg/dns/hosts/hosts_test.go
