+++ protokube/pkg/gossip/dns/hosts.go
@@ -18,7 +18,7 @@ package dns
 
 import (
 	"k8s.io/klog/v2"
-	"k8s.io/kops/protokube/pkg/gossip/dns/hosts"
+	"k8s.io/kops/pkg/dns/hosts"
 )
 
 // HostsFile stores DNS records into /etc/hosts
