+++ upup/pkg/fi/nodeup/nodetasks/update_etc_hosts_task.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	"k8s.io/klog/v2"
-	"k8s.io/kops/protokube/pkg/gossip/dns/hosts"
+	"k8s.io/kops/pkg/dns/hosts"
 	"k8s.io/kops/upup/pkg/fi"
 	"k8s.io/kops/upup/pkg/fi/nodeup/local"
 )
